# src/back/Imolt.Api/contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: ИМОЛТ — расчёт вывоза и утилизации отходов
  version: 1.2.0
  description: |
    Договор расчётной части сервиса ИМОЛТ: расчёт стоимости вывоза и
    утилизации строительных отходов по Москве и Московской области, подбор
    полигонов, маршрут, коммерческое предложение и заявка.

    **Как читать.** Операции сгруппированы по трём предметным областям
    расчётной части — «Справочники», «Расчёт», «Сделка». Границы областей
    объявлены записью ADR-0001, и группировка здесь — их внешняя проекция:
    по метке `x-область` видно, какая часть системы отвечает за операцию.

    **Что уже работает.** Отвечают только `/health` и `/ready`. Остальные
    операции помечены `x-состояние: объявлено`: форма запроса и ответа
    зафиксирована, сервер их пока не обслуживает. Мок поднимается из этого же
    файла, примеры ответов взяты из раздела 7 запроса на дизайн и согласованы
    между собой (адрес «Москва, ул. Годовикова, 9», дата актуальности
    17.09.2026).

    **Чего в договоре нет.** Поведение, не определённое заказчиком, не
    додумано: такие места помечены `x-вопросы` со ссылкой на открытый вопрос.
    Отложенные требования (R-047, R-062 — R-066) в версию 1 не входят.

    **Соглашения.**

    - Пути и имена полей — латиницей, тексты и описания — по-русски;
      идентификаторы кода остаются латиницей по правилам проекта.
    - Денежная сумма передаётся строкой с двумя знаками после точки: двоичная
...
      в макете (R-059).
  contact:
...
  license:
...
servers:
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/amount-conversions:
    post:
      operationId: convertAmounts
      tags: [Расчёт]
      summary: Пересчитать объём между тоннами и кубометрами
      description: |
        Подпись «≈ 7,5 т» под полем объёма. Пересчёт идёт на сервере, а не в
        браузере: коэффициент плотности — часть справочника, и второго места,
        где он применяется, быть не должно.

        По какому адресу определяется мера расчёта — по адресу вывоза или по
        полигону — не решено (Q-009), поэтому операция пересчитывает обе меры
        и меру расчёта сама не выбирает.
      x-область: расчёт
      x-требования: [R-014, R-015, R-016]
      x-вопросы: [Q-009]
      x-экран: [Э-02]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AmountConversionRequest' }
            example:
              items:
                - wasteGroupId: drevesina
                  quantity: { value: 15, unit: m3 }
      responses:
        '200':
          description: объём в обеих мерах
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AmountConversionResult' }
              example:
                items:
                  - wasteGroupId: drevesina
                    input: { value: 15, unit: m3 }
                    tons: 7.5
                    cubicMeters: 15
                    densityTonPerCubicMeter: 0.5
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/calculations:
    post:
      operationId: createCalculation
      tags: [Расчёт]
      summary: Рассчитать стоимость вывоза и утилизации
      description: |
        Главная операция сервиса. Принимает адрес вывоза с координатами, один
        или несколько типов отходов с объёмом и мерой, признак утилизации и
        фильтр расстояния. Возвращает по каждой группе отходов список
        полигонов со стоимостью перевозки, утилизации и итогом.

        Стоимость считается так: перевозка = объём в тоннах x цена группы за
        тонна-километр x расстояние по дорожной сети; утилизация = объём в
        тоннах x тариф полигона по этой группе (R-018). Ни коэффициенты, ни
        формула наружу не отдаются (R-058): клиент получает результат и его
        составляющие.

        Первая страница — 10 полигонов (R-029, R-060), остальные берутся
        операцией `listPlacementOptions`.

        Расчёт предварительный: поле `preliminary` истинно всегда, а
        допустимое отклонение финальной цены пока не задано (Q-010).
      x-область: расчёт
      x-требования: [R-002, R-003, R-018, R-019, R-020, R-021, R-022, R-058, R-059, R-060]
      x-вопросы: [Q-008, Q-010]
      x-экран: [Э-01, Э-02, Э-03, Э-12]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalculationRequest' }
            examples:
              twoWasteGroups:
                summary: путь UC-001 — бетон 20 т и древесина 15 м³
                value:
                  pickupAddress:
                    suggestionId: dadata-77-godovikova-9
                    value: г Москва, ул Годовикова, д 9
                    coordinates: { latitude: 55.8055, longitude: 37.6206 }
                  items:
                    - wasteGroupId: beton-lom
                      quantity: { value: 20, unit: t }
                    - wasteGroupId: drevesina
                      quantity: { value: 15, unit: m3 }
                  disposalRequired: true
                  distanceFilter: { mode: atMost, km: 60 }
      responses:
        '201':
          description: расчёт выполнен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Calculation' }
              examples:
                concreteResult:
                  $ref: '#/components/examples/ConcreteCalculation'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/AddressOutsideServiceArea' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
    get:
      operationId: listCalculations
      tags: [Расчёт]
      summary: Сохранённые расчёты кабинета
      description: Раздел «Расчёты» личного кабинета. Гостю недоступен.
      security:
        - bearerAuth: []
      x-область: расчёт
      x-требования: [R-049, R-008]
      x-вопросы: [Q-011]
      x-экран: [Э-10]
      x-состояние: реализовано
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: страница сохранённых расчётов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalculationSummaryPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/calculations/{calculationId}:
    parameters:
      - $ref: '#/components/parameters/CalculationIdPath'
    get:
      operationId: getCalculation
      tags: [Расчёт]
      summary: Расчёт целиком
      description: |
        Исходные данные, пересчитанные объёмы, выбор, распределение и
        актуальность данных. По этой операции страница восстанавливается по
        ссылке и после перезагрузки.
      x-область: расчёт
      x-требования: [R-002, R-019, R-048]
      x-экран: [Э-03, Э-04]
      x-состояние: реализовано
      responses:
        '200':
          description: расчёт
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Calculation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/calculations/{calculationId}/options:
    parameters:
      - $ref: '#/components/parameters/CalculationIdPath'
    get:
      operationId: listPlacementOptions
      tags: [Расчёт]
      summary: Варианты размещения по группе отходов
      description: |
        Список полигонов одной вкладки: сортировка, фильтр расстояния и
        «Показать ещё». По умолчанию список ранжирован по совокупной цене по
        возрастанию (R-024), фильтр — «до 50 км» (R-025, R-026).

        Пустой список — не ошибка: ответ несёт `emptyReason`, по которому
        интерфейс показывает «Нет полигонов … снимите фильтр» (Э-12).
      x-область: расчёт
      x-требования: [R-004, R-023, R-024, R-025, R-026, R-029, R-060]
      x-экран: [Э-03, Э-04, Э-12]
      x-состояние: реализовано
      parameters:
        - name: wasteGroupId
          in: query
          required: true
          description: вкладка группы отходов
          schema: { type: string }
        - name: sort
          in: query
          description: поле сортировки
          schema:
            type: string
            enum: [total, transport, disposal, distance]
            default: total
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: distanceMode
          in: query
          description: сторона фильтра расстояния
          schema:
            type: string
            enum: [atMost, atLeast]
            default: atMost
        - name: distanceKm
          in: query
          description: граница фильтра расстояния в километрах
          schema: { type: integer, minimum: 0, maximum: 1000, default: 50 }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: страница вариантов размещения
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlacementOptionPage' }
              examples:
                emptyByDistanceFilter:
                  summary: фильтр «не менее 60 км» не оставил полигонов
                  value:
                    items: []
                    total: 0
                    limit: 10
                    offset: 0
                    emptyReason: filteredOutByDistance
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/calculations/{calculationId}/selection:
    parameters:
      - $ref: '#/components/parameters/CalculationIdPath'
    put:
      operationId: setCalculationSelection
      tags: [Расчёт]
      summary: Выбрать полигоны и получить сводку
      description: |
        Флажки слева от названий полигонов. Выбор задаётся целиком, а не
        по одному: так сводка справа всегда отвечает тому, что отмечено.

        Выбор заблокированного полигона не запрещается — он остаётся за
        пользователем, но ответ несёт предупреждение (Э-12, R-028).
      x-область: расчёт
      x-требования: [R-027, R-028, R-032]
      x-экран: [Э-04, Э-12]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelectionRequest' }
            example:
              entries:
                - wasteGroupId: beton-lom
                  landfillId: vostok-timohovo
                - wasteGroupId: beton-lom
                  landfillId: iksha
      responses:
        '200':
          description: выбор сохранён, сводка пересчитана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectionState' }
              example:
                entries:
                  - wasteGroupId: beton-lom
                    landfillId: vostok-timohovo
                  - wasteGroupId: beton-lom
                    landfillId: iksha
                selectedLandfills: 2
                total: { amount: '39880.00', currency: RUB }
                warnings: []
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/calculations/{calculationId}/allocation:
    parameters:
      - $ref: '#/components/parameters/CalculationIdPath'
    put:
      operationId: setCalculationAllocation
      tags: [Расчёт]
      summary: Распределить объём между выбранными полигонами
      description: |
        Панель распределения: 20 т бетона делятся как 12 т и 8 т, итог
        пересчитывается по каждой части. Сумма частей обязана совпасть с
        объёмом группы — иначе операция отвечает 422 и ничего не меняет.
      x-область: расчёт
      x-требования: [R-030, R-032]
      x-экран: [Э-06]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationRequest' }
            example:
              entries:
                - wasteGroupId: beton-lom
                  landfillId: vostok-timohovo
                  quantity: { value: 12, unit: t }
                - wasteGroupId: beton-lom
                  landfillId: iksha
                  quantity: { value: 8, unit: t }
      responses:
        '200':
          description: распределение принято, итог пересчитан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationState' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/calculations/{calculationId}/route:
    parameters:
      - $ref: '#/components/parameters/CalculationIdPath'
    get:
      operationId: getCalculationRoute
      tags: [Расчёт]
      summary: Маршрут и сводка по выбранным полигонам
      description: |
        Всплывающее окно маршрута и панель «Получить маршрут»: расстояние,
        время в пути, ссылка на внешние карты и обременения по пути.

        Гостю детали маршрута закрыты; что именно закрыто подпиской, не
        решено (Q-011), поэтому ответ несёт поле `access`, а интерфейс по нему
        показывает замок вместо содержимого.

        Обременения (посты ДПС, весовой контроль) остаются пустыми, пока не
        назван источник данных и не подтверждена допустимость их показа
        (Q-003): поле в договоре есть, выдуманных значений в нём нет.
      x-область: расчёт
      x-требования: [R-005, R-032, R-033, R-034, R-035, R-050, R-057]
      x-вопросы: [Q-003, Q-008, Q-011]
      x-экран: [Э-05, Э-06]
      x-состояние: реализовано
      responses:
        '200':
          description: сводка маршрута по выбранным полигонам
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteSummary' }
              examples:
                guestWithoutSubscription:
                  summary: гость видит расстояние, но не детали маршрута
                  value:
                    access: { granted: false, reason: subscriptionRequired }
                    legs: []
                    total: { amount: '39880.00', currency: RUB }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Маркер сессии из операции `createSession`. Расчёт и коммерческое
        предложение задуманы доступными без входа; окончательный перечень
        закрытых операций — открытый вопрос Q-011.
  parameters:
...
    Limit:
      name: limit
      in: query
      description: |
        Сколько записей вернуть. По умолчанию 10: первый экран результатов
        показывает десять полигонов, остальные догружаются (R-029, R-060).
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    Offset:
      name: offset
      in: query
      description: сколько записей пропустить
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
...
    CalculationIdPath:
      name: calculationId
      in: path
      required: true
      description: идентификатор расчёта
      schema: { type: string }
      example: 5f0a1c2b-77d9-4f61-9a6f-9c3d1b2a8e40
  responses:
    BadRequest:
      description: запрос не разобран или параметр вне допустимых значений
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:validation'
            title: Запрос не прошёл проверку
            status: 400
            detail: Параметр limit должен быть от 1 до 100
    NotFound:
      description: запись не найдена
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:not-found'
            title: Запись не найдена
            status: 404
    Unauthorized:
      description: операция требует входа
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:authentication-required'
            title: Нужен вход
            status: 401
...
    UnprocessableContent:
      description: запрос разобран, но нарушает правило предметной области
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:allocation-mismatch'
            title: Распределение не сходится с объёмом
            status: 422
            detail: По группе «Лом бетона» распределено 18 т из 20 т
            errors:
              - field: entries[1].quantity.value
                message: увеличьте на 2 т либо уменьшите объём группы
    AddressOutsideServiceArea:
      description: |
        Адрес вывоза вне Москвы и Московской области либо без координат.
        Расчёт не запускается, поле подсвечивается (R-012).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:address-outside-service-area'
            title: Адрес вне зоны обслуживания
            status: 422
            detail: Сервис считает вывоз по Москве и Московской области
    UpstreamUnavailable:
      description: |
        Внешняя служба расстояний или подсказок не ответила либо исчерпан
        лимит обращений (риск AR-007). Сохранённые данные при этом остаются
        доступны: отказ источника не равен отказу обслуживания.
      headers:
        Retry-After:
          description: через сколько секунд повторить
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:distance-service-unavailable'
            title: Не удалось рассчитать расстояния
            status: 503
            detail: Повторите попытку
...
  examples:
    ConcreteCalculation:
      summary: расчёт UC-001 — первая страница по группе «Лом бетона»
      value:
        id: 5f0a1c2b-77d9-4f61-9a6f-9c3d1b2a8e40
        createdAt: '2026-09-17T12:00:00+03:00'
        preliminary: true
        pickupAddress:
          value: г Москва, ул Годовикова, д 9
          coordinates: { latitude: 55.8055, longitude: 37.6206 }
          area: moscow
        disposalRequired: true
        distanceFilter: { mode: atMost, km: 60 }
        dataFreshness:
          pricesUpdatedAt: '2026-09-17'
          statusesUpdatedAt: '2026-09-17'
          landfillsWithStaleData: 1
        items:
          - wasteGroupId: beton-lom
            wasteGroupName: Лом бетона и железобетона
            input: { value: 20, unit: t }
            tons: 20
          - wasteGroupId: drevesina
            wasteGroupName: Древесина от разборки
            input: { value: 15, unit: m3 }
            tons: 7.5
        results:
          - wasteGroupId: beton-lom
            options:
              items:
                - landfillId: vostok-timohovo
                  landfillName: Комплекс переработки «Восток»
                  address: Московская обл., Богородский г. о., д. Тимохово
                  distanceKm: 45
                  transportCost: { amount: '10800.00', currency: RUB }
                  disposalCost: { amount: '9000.00', currency: RUB }
                  totalCost: { amount: '19800.00', currency: RUB }
                  status: active
                  statusUpdatedAt: '2026-09-17'
                - landfillId: iksha
                  landfillName: Площадка «Икша»
                  address: Московская обл., Дмитровский г. о., пос. Икша
                  distanceKm: 52
                  transportCost: { amount: '12480.00', currency: RUB }
                  disposalCost: { amount: '7600.00', currency: RUB }
                  totalCost: { amount: '20080.00', currency: RUB }
                  status: active
                  statusUpdatedAt: '2026-09-17'
              total: 5
              limit: 10
              offset: 0
              emptyReason: null
  schemas:
...
    Unit:
      type: string
      description: мера объёма отходов — тонны либо кубические метры
      enum: [t, m3]
    Quantity:
      type: object
      required: [value, unit]
      properties:
        value: { type: number, format: double, minimum: 0, example: 20 }
        unit: { $ref: '#/components/schemas/Unit' }
      additionalProperties: false
...
    Problem:
      type: object
      description: |
        Описание ошибки по RFC 9457. Поле `type` — устойчивый код причины:
        интерфейс ветвится по нему, а не по тексту заголовка.
      required: [type, title, status]
      properties:
        type: { type: string, example: 'urn:imolt:problem:validation' }
        title: { type: string, example: Запрос не прошёл проверку }
        status: { type: integer, example: 400 }
        detail: { type: string }
        instance: { type: string, description: 'путь запроса, на котором возникла ошибка' }
        errors:
          type: array
          description: разбор по полям — для форм
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    PickupAddress:
      type: object
      description: |
        Адрес вывоза. Принимается только выбранный из подсказок: расчёт
        опирается на координаты, а не на набранную строку.
      required: [value, coordinates]
      properties:
        suggestionId: { type: string }
        value: { type: string }
        coordinates: { $ref: '#/components/schemas/Coordinates' }
        area: { $ref: '#/components/schemas/ServiceArea' }
...
    AmountConversionRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [wasteGroupId, quantity]
            properties:
              wasteGroupId: { type: string }
              quantity: { $ref: '#/components/schemas/Quantity' }
    AmountConversionResult:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [wasteGroupId, input, tons, cubicMeters, densityTonPerCubicMeter]
            properties:
              wasteGroupId: { type: string }
              input: { $ref: '#/components/schemas/Quantity' }
              tons: { type: number, format: double }
              cubicMeters: { type: number, format: double }
              densityTonPerCubicMeter: { type: number, format: double }
    DistanceFilter:
      type: object
      description: |
        Фильтр расстояния «до N км» либо «не менее N км». По умолчанию
        «до 50 км» (R-025, R-026).
      required: [mode, km]
      properties:
        mode: { type: string, enum: [atMost, atLeast], default: atMost }
        km: { type: integer, minimum: 0, maximum: 1000, default: 50 }
      additionalProperties: false
    CalculationRequest:
      type: object
      required: [pickupAddress, items]
      properties:
        pickupAddress: { $ref: '#/components/schemas/PickupAddress' }
        items:
          type: array
          minItems: 1
          maxItems: 20
          description: типы отходов расчёта; в одном расчёте их может быть несколько
          items:
            type: object
            required: [wasteGroupId, quantity]
            properties:
              wasteGroupId: { type: string }
              quantity: { $ref: '#/components/schemas/Quantity' }
        disposalRequired:
          type: boolean
          default: true
          description: |
            Флажок «Нужна утилизация на полигоне». Когда ложно, совокупная
            цена состоит только из перевозки (R-021).
        distanceFilter: { $ref: '#/components/schemas/DistanceFilter' }
      additionalProperties: false
    CalculationItem:
      type: object
      required: [wasteGroupId, wasteGroupName, input, tons]
      properties:
        wasteGroupId: { type: string }
        wasteGroupName: { type: string }
        input: { $ref: '#/components/schemas/Quantity' }
        tons:
          type: number
          format: double
          description: объём в тоннах — мера, в которой считается стоимость
...
    PlacementOptionPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/PlacementOption' }
            emptyReason:
              type: string
              nullable: true
              description: |
                Почему список пуст. `noLandfillsForWasteGroup` — эту группу не
                принимает ни один полигон; `filteredOutByDistance` — все
                отсеяны фильтром расстояния. Пустой список ошибкой не является.
              enum: [noLandfillsForWasteGroup, filteredOutByDistance, null]
    Calculation:
      type: object
      description: |
        Расчёт целиком: исходные данные, пересчитанные объёмы, результаты по
        каждой группе отходов, выбор и распределение. Хранит дату
        актуальности данных, на которых посчитан (R-048).
      required: [id, createdAt, preliminary, pickupAddress, disposalRequired, items, results, dataFreshness]
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        preliminary:
          type: boolean
          description: |
            Расчёт предварительный всегда (R-059). Допустимое отклонение
            финальной цены заказчиком не названо — Q-010.
        pickupAddress: { $ref: '#/components/schemas/PickupAddress' }
        disposalRequired: { type: boolean }
        distanceFilter: { $ref: '#/components/schemas/DistanceFilter' }
        items:
          type: array
          items: { $ref: '#/components/schemas/CalculationItem' }
        results:
          type: array
          description: по одной вкладке на группу отходов
          items:
            type: object
            required: [wasteGroupId, options]
            properties:
              wasteGroupId: { type: string }
              options: { $ref: '#/components/schemas/PlacementOptionPage' }
        selection: { $ref: '#/components/schemas/SelectionState' }
        allocation: { $ref: '#/components/schemas/AllocationState' }
        dataFreshness: { $ref: '#/components/schemas/DataFreshness' }
...
    CalculationSummaryPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/CalculationSummary' }
...
    SelectionRequest:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          description: выбор задаётся целиком; пустой массив снимает выбор
          items: { $ref: '#/components/schemas/SelectionEntry' }
      additionalProperties: false
    SelectionState:
      type: object
      required: [entries, selectedLandfills, total, warnings]
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/SelectionEntry' }
        selectedLandfills: { type: integer }
        total: { $ref: '#/components/schemas/Money' }
        warnings:
          type: array
          description: |
            Предупреждения выбора. Заблокированный полигон остаётся выбранным
            по решению пользователя — предупреждение не отменяет выбор (Э-12).
          items:
            type: object
            required: [code, landfillId, message]
            properties:
              code: { type: string, enum: [landfillBlocked, landfillDataStale] }
              landfillId: { type: string }
              message: { type: string }
...
    AllocationRequest:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/AllocationEntry' }
      additionalProperties: false
    AllocationState:
      type: object
      required: [entries, total]
      properties:
        entries:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/AllocationEntry'
              - type: object
                required: [transportCost, totalCost]
                properties:
                  transportCost: { $ref: '#/components/schemas/Money' }
                  disposalCost:
                    allOf: [{ $ref: '#/components/schemas/Money' }]
                    nullable: true
                  totalCost: { $ref: '#/components/schemas/Money' }
        total: { $ref: '#/components/schemas/Money' }
...
    RouteSummary:
      type: object
      required: [access, legs, total]
      properties:
        access: { $ref: '#/components/schemas/Access' }
        legs:
          type: array
          items: { $ref: '#/components/schemas/RouteLeg' }
        total: { $ref: '#/components/schemas/Money' }
...